einhalten von Model View Controler
easy fix
