+++ ControlDeveloper2/src/GUI/Panels/PanelMenuBar.java
@@ -5,13 +5,11 @@ import java.awt.event.ActionListener;
 import java.io.File;
 import java.util.Vector;
 
-import javax.swing.Icon;
 import javax.swing.JFileChooser;
 import javax.swing.JMenu;
 import javax.swing.JMenuBar;
 import javax.swing.JMenuItem;
 import javax.swing.JOptionPane;
-import javax.swing.JPanel;
 
 import Controller.ControlDevelepor;
 import GUI.ViewControlDevelepor;
